requirements.txt
# 电路图检索系统依赖

# Web 应用
gradio>=4.0.0

# 核心依赖
pandas>=2.0.0
numpy>=1.24.0

# BGE Reranker
sentence-transformers>=2.2.0
torch>=2.0.0

# 中文分词
jieba>=0.42.0

# HTTP 请求（用于 LLM API）
requests>=2.28.0

# 环境变量管理
python-dotenv>=1.0.0
